html: split worker global-scope interfaces into their own modules
Give each worker interface its own module under content/src/html,
mirroring the spec section it is specified in (workers.html §10.2.1):
worker.rs keeps the Worker interface and WorkerType, the
WorkerGlobalScope platform object moves to worker_global_scope.rs,
and dedicated_worker_global_scope.rs holds the DedicatedWorkerGlobalScope
interface marker (whose dedicated members the dedicated realm's
WorkerGlobalScope platform object implements). Mirror the split in the
bindings layer: dedicated_worker_global_scope.rs now carries the
DedicatedWorkerGlobalScope WebIdlInterface and the shared worker
global-scope downcast/event-handler helpers stay in worker_global_scope.rs
as pub(crate). Fix the DedicatedWorkerGlobalScope marker's anchor to the
real spec id (#dedicatedworkerglobalscope). Bodies are moved verbatim;
no behavior change (WPT: 217 executed, 0 unexpected; TLA+ specs pass).
