+++ content/src/html.rs
@@ -9,6 +9,7 @@ mod activation_behavior;
 pub(crate) mod channel_messaging;
 pub(crate) use channel_messaging::ChannelMessaging;
 pub(crate) mod dedicated_worker_agent;
+pub(crate) mod dedicated_worker_global_scope;
 pub(crate) mod dispatch;
 pub(crate) mod environment_settings_object;
 pub(crate) mod event_handler;
@@ -34,6 +35,7 @@ mod window;
 mod window_or_worker_global_scope;
 pub(crate) mod windowproxy;
 pub(crate) mod worker;
+pub(crate) mod worker_global_scope;
 mod worker_location;
 mod worker_navigator;
 
@@ -51,6 +53,7 @@ pub use global_scope::GlobalScope;
 pub use global_scope::GlobalScopeKind;
 pub(crate) use global_scope::TimerHandler;
 
+pub(crate) use dedicated_worker_global_scope::DedicatedWorkerGlobalScope;
 pub use html_anchor_element::HTMLAnchorElement;
 pub(crate) use html_dom_tree::{
     run_dom_post_connection_steps_for_document, run_dom_removing_steps_for_document,
@@ -79,7 +82,8 @@ pub(crate) use window::window_computed_style_properties_for_element;
 pub(crate) use window::{PostMessageOptions, window_post_message_steps};
 pub(crate) use window_or_worker_global_scope::WindowOrWorkerGlobalScope;
 pub(crate) use windowproxy::WindowProxy;
-pub(crate) use worker::{DedicatedWorkerGlobalScope, Worker, WorkerGlobalScope, WorkerType};
+pub(crate) use worker::{Worker, WorkerType};
+pub(crate) use worker_global_scope::WorkerGlobalScope;
 pub(crate) use worker_location::WorkerLocation;
 pub(crate) use worker_navigator::WorkerNavigator;
