+++ content/src/html/dedicated_worker_global_scope.rs
@@ -1,0 +1,5 @@
+use js_engine::gc_struct;
+
+/// <https://html.spec.whatwg.org/#dedicatedworkerglobalscope>
+#[gc_struct]
+pub(crate) struct DedicatedWorkerGlobalScope {}
